requirements.txt
absl-py==0.2.2
appnope==0.1.0
astor==0.6.2
bleach==1.5.0
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.3
dask==0.17.5
decorator==4.1.2
entrypoints==0.2.3
gast==0.2.0
gevent==1.3.4
greenlet==0.4.13
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
image==1.5.24
image2pipe==0.1.7
imgaug==0.2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.1.6
keras-vis==0.4.1
kiwisolver==1.0.1
llvmlite==0.23.2
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
networkx==2.1
nose==1.3.7
notebook==5.0.0
numba==0.38.1
numpy==1.14.5
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
RXPY==0.1.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.10.0
tblib==1.3.2
tensorboard==1.8.0
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
toolz==0.9.0 
torchvision==0.2.1
tornado==4.5.1
tqdm==4.19.5
traitlets==4.3.2
umap-learn==0.2.5
wcwidth==0.1.7
webencodings==0.5.1
websocket==0.2.1
Werkzeug==0.14.1
widgetsnbextension==2.0.0
